+++ src/log2.c
@@ -27,7 +27,7 @@ char* log_2_binary(int d, int num_places)
 {
     int base = 2;
     int c = 1;
-    int (*p)(int) = &log_2_binary_p;
+    long long int (*p)(long long int) = &log_2_binary_p;
     long double result = compute_bbp(d, base, c, p, false);
 
     return convert_floating_decimal_to_base(result, num_places, base);
@@ -42,7 +42,7 @@ char* log_2_binary(int d, int num_places)
  *
  *  returns: p(k)
  */
-int log_2_binary_p(int k)
+long long int log_2_binary_p(long long int k)
 {
     return k;
 }
